+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Types of changes include:
 
 ## Upcoming/Unreleased
 
+- NEW: There is now a link to the help site in the footer on the homepage and other logged-out pages.
+- CHANGED: The links to the About, Privacy, and Contact pages have been returned to the user menu.
+
 ## 1.0.6
 
 - CHANGED: The _Set new project total?_ toggle on the Add Progress dialog will now save the last setting you used. (h/t Courtney)